[automated] Regenerate samples with new tracing.
