+++ samples/gen-go-db-only/tracing/tracing.go
@@ -12,10 +12,11 @@ import (
 
 	"go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+	"go.opentelemetry.io/contrib/propagators/aws/xray"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/exporters/otlp"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpgrpc"
 	"go.opentelemetry.io/otel/propagation"
+	sdkexporttrace "go.opentelemetry.io/otel/sdk/trace"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 	"go.opentelemetry.io/otel/trace"
@@ -23,7 +24,10 @@ import (
 )
 
 // propagator to use.
-var propagator propagation.TextMapPropagator = propagation.TraceContext{} // traceparent header
+var propagator propagation.TextMapPropagator = propagation.NewCompositeTextMapPropagator(
+	propagation.TraceContext{}, // traceparent header
+	xray.Propagator{},          // x-amzn-trace-id header
+)
 
 // defaultCollectorPort was changed from 55860 in November and the Go library
 // hasn't been updated when it is updated we can use otlp.DefaultCollectorPort
@@ -32,7 +36,7 @@ var defaultCollectorPort uint16 = 4317
 // SetupGlobalTraceProviderAndExporter sets up an exporter to export,
 // as well as the opentelemetry global trace provider for trace generators to use.
 // The exporter and provider are returned in order for the caller to defer shutdown.
-func SetupGlobalTraceProviderAndExporter(ctx context.Context) (sdktrace.SpanExporter, *sdktrace.TracerProvider, error) {
+func SetupGlobalTraceProviderAndExporter() (sdkexporttrace.SpanExporter, *sdktrace.TracerProvider, error) {
 	// 1. set up exporter
 	// 2. set up tracer provider
 	// 3. assign global tracer provider
@@ -46,25 +50,20 @@ func SetupGlobalTraceProviderAndExporter(ctx context.Context) (sdktrace.SpanExpo
 	} else if v := os.Getenv("TRACING_SAMPLING_PROBABILITY"); v != "" {
 		samplingProbabilityFromEnv, err := strconv.ParseFloat(v, 64)
 		if err != nil {
-			return nil, nil, fmt.Errorf("could not parse '%s' to float", v)
+			return nil, nil, fmt.Errorf("could not parse '%s' to integer", v)
 		}
 		samplingProbability = samplingProbabilityFromEnv
 	}
-
-	addr := fmt.Sprintf("%s:%d", otlp.DefaultCollectorHost, defaultCollectorPort)
 
 	// Every 15 seconds we'll try to connect to opentelemetry collector at
 	// the default location of localhost:4317
 	// When running in production this is a sidecar, and when running
 	// locally this is a locally running opetelemetry-collector.
-	driver := otlpgrpc.NewDriver(
-		otlpgrpc.WithReconnectionPeriod(15*time.Second),
-		otlpgrpc.WithEndpoint(addr),
-		otlpgrpc.WithInsecure(),
-	)
 	exporter, err := otlp.NewExporter(
-		ctx,
-		driver,
+		context.Background(),
+		otlp.WithAddress(fmt.Sprintf("%s:%d", otlp.DefaultCollectorHost, defaultCollectorPort)),
+		otlp.WithReconnectionPeriod(15*time.Second),
+		otlp.WithInsecure(),
 	)
 	if err != nil {
 		return nil, nil, fmt.Errorf("error creating exporter: %v", err)
@@ -72,10 +71,6 @@ func SetupGlobalTraceProviderAndExporter(ctx context.Context) (sdktrace.SpanExpo
 
 	tp := newTracerProvider(exporter, samplingProbability)
 	otel.SetTracerProvider(tp)
-	logger.FromContext(ctx).InfoD("starting-tracer", logger.M{
-		"address":       addr,
-		"sampling-rate": samplingProbability,
-	})
 	return exporter, tp, nil
 }
 
@@ -89,17 +84,14 @@ func SetupGlobalTraceProviderAndExporterForTest() (*tracetest.InMemoryExporter, 
 	return exporter, tp, nil
 }
 
-func newTracerProvider(exporter sdktrace.SpanExporter, samplingProbability float64) *sdktrace.TracerProvider {
+func newTracerProvider(exporter sdkexporttrace.SpanExporter, samplingProbability float64) *sdktrace.TracerProvider {
 	return sdktrace.NewTracerProvider(
-		// We use the default ID generator. In order for sampling to work (at least with this sampler)
-		// the ID generator must generate trace IDs uniformly at random from the entire space of uint64.
-		// For example, the default x-ray ID generator does not do this.
-		sdktrace.WithSampler(sdktrace.ParentBased(sdktrace.TraceIDRatioBased(samplingProbability))),
-		// These maximums are to guard against something going wrong and sending a ton of data unexpectedly
-		sdktrace.WithSpanLimits(sdktrace.SpanLimits{
-			AttributeCountLimit: 100,
-			EventCountLimit:     100,
-			LinkCountLimit:      100,
+		sdktrace.WithConfig(sdktrace.Config{
+			IDGenerator:          xray.NewIDGenerator(),
+			DefaultSampler:       sdktrace.ParentBased(sdktrace.TraceIDRatioBased(samplingProbability)),
+			MaxEventsPerSpan:     100,
+			MaxAttributesPerSpan: 100,
+			MaxLinksPerSpan:      100,
 		}),
 		sdktrace.WithSyncer(exporter),
 	)
@@ -107,7 +99,6 @@ func newTracerProvider(exporter sdktrace.SpanExporter, samplingProbability float
 
 // MuxServerMiddleware returns middleware that should be attached to a gorilla/mux server.
 // It does two things: starts spans, and adds span/trace info to the request-specific logger.
-// Right now we only support logging IDs in the format that Datadog expects.
 func MuxServerMiddleware(serviceName string) func(http.Handler) http.Handler {
 	otlmux := otelmux.Middleware(serviceName, otelmux.WithPropagators(propagator))
 	return func(h http.Handler) http.Handler {
@@ -115,7 +106,9 @@ func MuxServerMiddleware(serviceName string) func(http.Handler) http.Handler {
 			// otelmux has extracted the span. now put it into the ctx-specific logger
 			s := trace.SpanFromContext(r.Context())
 			if sc := s.SpanContext(); sc.HasTraceID() {
-				spanID, traceID := sc.SpanID().String(), sc.TraceID().String()
+				spanID, traceID := sc.SpanID.String(), sc.TraceID.String()
+				// be everything to everyone
+				// datadog:
 				// datadog converts hex strings to uint64 IDs, so log those so that correlating logs and traces works
 				if len(traceID) == 32 && len(spanID) == 16 { // opentelemetry format: 16 byte (32-char hex), 8 byte (16-char hex) trace and span ids
 					traceIDBs, _ := hex.DecodeString(traceID)
@@ -125,6 +118,11 @@ func MuxServerMiddleware(serviceName string) func(http.Handler) http.Handler {
 					logger.FromContext(r.Context()).AddContext("span_id",
 						fmt.Sprintf("%d", binary.BigEndian.Uint64(spanIDBs)))
 				}
+				// newrelic:
+				logger.FromContext(r.Context()).AddContext("span.id", spanID)
+				logger.FromContext(r.Context()).AddContext("trace.id", traceID)
+				// x-ray
+				logger.FromContext(r.Context()).AddContext("x-ray", fmt.Sprintf("1-%s-%s", spanID, traceID))
 			}
 			h.ServeHTTP(rw, r)
 		}))
@@ -164,8 +162,8 @@ func (rt roundTripper) RoundTrip(r *http.Request) (*http.Response, error) {
 func ExtractSpanAndTraceID(r *http.Request) (traceID, spanID string) {
 	s := trace.SpanFromContext(r.Context())
 	if s.SpanContext().HasTraceID() {
-		return s.SpanContext().TraceID().String(), s.SpanContext().SpanID().String()
+		return s.SpanContext().TraceID.String(), s.SpanContext().SpanID.String()
 	}
-	sc := trace.SpanContextFromContext(propagator.Extract(r.Context(), propagation.HeaderCarrier(r.Header)))
-	return sc.SpanID().String(), sc.TraceID().String()
+	sc := trace.RemoteSpanContextFromContext(propagator.Extract(r.Context(), r.Header))
+	return sc.SpanID.String(), sc.TraceID.String()
 }
